finish the assainment

+++ src/Pages/MyBooking/MyBooking.jsx
@@ -5,6 +5,7 @@ import moment from "moment";
 import { AuthContext } from "../../Provider/AuthProvider/AuthProvider";
 import axios from "axios";
 import { Helmet } from "react-helmet";
+import Swal from "sweetalert2";
 
 
 const MyBooking = () => {
@@ -31,19 +32,58 @@ const MyBooking = () => {
     const now = moment();
     var formattedNow = now.format("YYYY-MM-DD HH:mm:ss");
     const validDate = {formattedNow}
-    fetch(`https://assainment-11-server.vercel.app/api/v1/bookingData/${_id}`,{
-        method: 'DELETE',
-        headers:{
-          'content-type': 'application/json'
-        },
-        body: JSON.stringify(validDate)
-    })
-    .then(res=>res.json())
-    .then(data=>{
-        console.log(data);
-        const remaning = myBooking.filter(remaningData=> remaningData._id !== _id)
-        setMyBooking(remaning)
-    })
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!"
+    }).then((result) => {
+      if (result.isConfirmed) {
+
+        fetch(`https://assainment-11-server.vercel.app/api/v1/bookingData/${_id}`,{
+          method: 'DELETE',
+          headers:{
+            'content-type': 'application/json'
+          },
+          body: JSON.stringify(validDate)
+      })
+      .then(res=>res.json())
+      .then(data=>{
+          console.log(data);
+          const remaning = myBooking.filter(remaningData=> remaningData._id !== _id)
+          setMyBooking(remaning)
+          Swal.fire({
+            title: "Deleted!",
+            text: "Your file has been deleted.",
+            icon: "success"
+          });
+      })
+
+
+      
+      }
+    });
+
+
+
+
+
+    // fetch(`https://assainment-11-server.vercel.app/api/v1/bookingData/${_id}`,{
+    //     method: 'DELETE',
+    //     headers:{
+    //       'content-type': 'application/json'
+    //     },
+    //     body: JSON.stringify(validDate)
+    // })
+    // .then(res=>res.json())
+    // .then(data=>{
+    //     console.log(data);
+    //     const remaning = myBooking.filter(remaningData=> remaningData._id !== _id)
+    //     setMyBooking(remaning)
+    // })
 }
     return (
       <div className="mt-10 space-y-4 mb-10 w-[1200px] mx-auto  ">
